hdmitx21: Optimization of standby power consumption [1/1]
PD#SWPL-161016

Problem:
When s7 is in standby and wakes up, it will turn off and on the vpu
power domain. When it is turned off, the reg of the relevant modules
will be reset, the output no signal when it wakes up.

Solution:
hdmitx driver needs to reinitialize the required top register when
it wake up

Verify:
s7

Test:
DRM-TX-78, DRM-TX-79, DRM-TX-80, DRM-TX-81

Change-Id: Ie9d458087fcc9ea672ec477d46b88a6d7ddcec81
Signed-off-by: xiang.wu1 <xiang.wu1@amlogic.com>