nand: spinand bitflip threshold set to 75% ecc strength [1/1]
PD#SWPL-162429

Problem:
It is not really very good for ecc threshold
to be equal to the maximum capacity of ecc.
we refer to raw nand method to set it.

Solution:
Refer to raw nand method (nand_base.c) to set ecc
threshold to the 75% of ecc maximum capacity.

Verify:
S1A-BG201

Change-Id: I92fea91008a93fd5472bfcadcd338e89e985d40c
Signed-off-by: zhikui.cui <zhikui.cui@amlogic.com>