audio: fix earc Cert HFR5-1-35 failed [1/1]
PD#SWPL-159771

Problem:
get CDS info which is not ready by the spec.
rootcause:when the earc connection success, it send
the earc state to framework, then send connet msg to
audio hal, and audio hal get cds info, but the cds is
not ready by the spec.

Solution:
when we get the cds ready info, then send the earc
connection success event.

Verify:
T3X

Change-Id: Ie80a0523d71330d7dd47040cc7f7ec57228f3133
Signed-off-by: jian zhou <jian.zhou@amlogic.com>